requirements.txt
discord.py==1.7.3
openai==0.27.0
python-dotenv==0.19.2
farm-haystack